add user role

+++ src/main/java/com/anhngo/nhaichuttruyen/repository/RolesRepository.java
@@ -1,9 +1,10 @@
 package com.anhngo.nhaichuttruyen.repository;
 
 import com.anhngo.nhaichuttruyen.entity.Roles;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolesRepository extends JpaRepository<Roles, Long> {
+public interface RolesRepository extends JpaRepository<Roles, Integer> {
 }
